requirements.txt
discord.py
matplotlib
numpy